+++ src/main/java/com/example/uambite/dto/response/DetallePedidoResponse.java
@@ -5,6 +5,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.UUID;
 
@@ -15,8 +16,8 @@ import java.util.UUID;
 public class DetallePedidoResponse {
     private UUID id;
     private Integer cantidad;
-    private Double precioUnitario;
-    private Double subtotal;
+    private BigDecimal precioUnitario;
+    private BigDecimal subtotal;
     private String producto;
     private UUID productoId;
     private List<DetallePedidoIngredienteExtraResponse> ingredientesExtra;
